pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.3.8",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.11.1",
]

[dependency-groups]
dev = [
    "codespell>=2.4.1",
    "pyright>=1.1.402",
    "ruff>=0.12.1",
]
testing = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
]
